Create Delete the downloaded files
In an application, there may be instances in which we download the files with some extension types. In our case , it is .exe files extension.
we write a common utility function in Utility.cs in our page object pattern in which we uses the "Directory.GetFiles(string downloadPath,string extension).